fix: folders' structure conflict
