+++ src/api/package.json
@@ -10,18 +10,13 @@
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "express": "~4.16.1",
-    "firebase-functions": "^5.0.1",
-    "google-libphonenumber": "^3.2.38",
     "hara-pay.application": "file:../application",
     "http-errors": "~1.6.3",
     "moment": "^2.30.1",
     "morgan": "~1.9.1",
     "pug": "2.0.0-beta11",
-    "randomstring": "^1.3.0",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-express": "^5.0.1",
-    "tinyurl": "^1.1.7",
-    "ussd-router": "^0.1.6"
+    "swagger-ui-express": "^5.0.1"
   },
   "devDependencies": {
     "nodemon": "^3.1.4"
