+++ src/application/package.json
@@ -5,6 +5,12 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "@hashgraph/sdk": "^2.49.2"
+    "@hashgraph/sdk": "^2.49.2",
+    "firebase-functions": "^5.0.1",
+    "google-libphonenumber": "^3.2.38",
+    "moment": "^2.30.1",
+    "randomstring": "^1.3.0",
+    "tinyurl": "^1.1.7",
+    "ussd-router": "^0.1.6"
   }
 }
